rename modal to model
