+++ api/models/__init__.py
@@ -1,6 +1,6 @@
 from .patient_model import Patient
 from .visit_model import Visit
-from .vitals_modal import Vitals
+from .vitals_model import Vitals
 from .consult_model import Consult
 from .diagnosis_model import Diagnosis
 from .medication_review_model import MedicationReview
