+++ api/models/vitals_model.py
